requirements.txt
# FastAPI and ASGI Server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Google Cloud Platform
google-cloud-secret-manager==2.21.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-auth==2.36.0

# Data Validation and Settings
pydantic==2.10.3
pydantic-settings==2.7.0

# Utilities
python-multipart==0.0.20
